# PA: overall performance improvements