- re-use values that need to be calcualted (instead of calculating them again)
- store re-usable information in "itemData" table
- extract variable declaration outside for-loop

Noticable difference in ESO Profiler when accessing the bank for the first time:

![image](https://user-images.githubusercontent.com/14639044/97804715-f697bc00-1c51-11eb-9d60-5e52480d65bc.png)

| Function | Time Before  | Time After | Time Change | Calls Before  | Calls After | CallsChange | 
| ------------- | ------------- | ------------- | ------------- |------------- |------------- |------------- |
| < anonymous function > | 218ms | 30ms | 626% faster | 8 calls | 8 calls | - |
| depositOrWithdrawCustomItems | 202ms | 23ms | 778% faster | 1 call | 1 call | - |
| getPAIdemLinkIdentifier | 176ms | 21ms + 19ms | 340% faster | 3118 calls | 544+294 calls | 277% less calls |
| isValueInTable | 1.493ms | 0.023ms | 6391% faster | 1725 calls | 33 calls | 5127% less calls |
| _getCraftingTypeFromWritItemLink | 1.446ms | 0.055ms | 2529% faster | 1869 calls | 130 calls | 1337% less calls |

An additional performance improvement of up to 200% can be expected for subsequent bank visits when cached data can be utilized.